Rename Azure package to *.key.vault.azure
