+++ key-vault/azure-key-vault/src/main/java/com/quorum/tessera/key/vault/azure/AzureCredentialNotSetException.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.key.vault;
+package com.quorum.tessera.key.vault.azure;
 
 public class AzureCredentialNotSetException extends IllegalStateException {
