+++ key-vault/azure-key-vault/src/test/java/com/quorum/tessera/key/vault/azure/AzureCredentialNotSetExceptionTest.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.key.vault;
+package com.quorum.tessera.key.vault.azure;
 
 import org.junit.Test;
