+++ key-vault/azure-key-vault/src/test/java/com/quorum/tessera/key/vault/azure/AzureKeyVaultServiceTest.java
@@ -1,8 +1,11 @@
-package com.quorum.tessera.key.vault;
+package com.quorum.tessera.key.vault.azure;
 
 import com.microsoft.azure.keyvault.models.SecretBundle;
 import com.microsoft.azure.keyvault.requests.SetSecretRequest;
 import com.quorum.tessera.config.KeyVaultConfig;
+import com.quorum.tessera.key.vault.VaultSecretNotFoundException;
+import com.quorum.tessera.key.vault.azure.AzureKeyVaultClientDelegate;
+import com.quorum.tessera.key.vault.azure.AzureKeyVaultService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.ArgumentCaptor;
